backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
langchain==0.1.9
langchain-community==0.0.24
langchain-google-genai==0.0.9
faiss-cpu==1.7.4
sentence-transformers==2.5.1
pydantic>=2.7.0
python-multipart==0.0.9
deep-translator==1.11.4
